# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "status",
                        "mfaEnrollmentRequired",
                        "challengeExpiresAt"
                      ],
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "mfa_required"
                          ]
                        },
                        "mfaEnrollmentRequired": {
                          "type": "boolean"
                        },
                        "challengeExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "username",
                        "displayName",
                        "role",
                        "status",
                        "mfaEnabled",
                        "mustChangePassword"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "editor",
                            "reviewer",
                            "publisher",
                            "system_admin"
                          ]
                        },
                        "status": {
                          "type": "string"
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        },
                        "mustChangePassword": {
                          "type": "boolean"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "operationId": "rotateCsrf",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "csrfToken"
                      ],
                      "properties": {
                        "csrfToken": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "username",
                        "displayName",
                        "role",
                        "status",
                        "mfaEnabled",
                        "mustChangePassword"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "editor",
                            "reviewer",
                            "publisher",
                            "system_admin"
                          ]
                        },
                        "status": {
                          "type": "string"
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        },
                        "mustChangePassword": {
                          "type": "boolean"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "status",
                        "recoveryAction"
                      ],
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "logged_out"
                          ]
                        },
                        "recoveryAction": {
                          "type": "string",
                          "enum": [
                            "delete"
                          ]
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "email",
                          "username",
                          "displayName",
                          "role",
                          "status",
                          "mfaEnabled",
                          "mustChangePassword"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "username": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "editor",
                              "reviewer",
                              "publisher",
                              "system_admin"
                            ]
                          },
                          "status": {
                            "type": "string"
                          },
                          "mfaEnabled": {
                            "type": "boolean"
                          },
                          "mustChangePassword": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-users"
        ]
      },
      "post": {
        "operationId": "createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/api/v1/admin/invites": {
      "post": {
        "operationId": "createInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-users"
        ]
      }
    },
...
  },
  "info": {
    "title": "DanangMap API",
    "description": "DanangMap v2 spatial CMS contract",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-danangmap_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "example": "editor@example.gov.vn"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "default": "totp"
          },
          "code": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "method",
          "code"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "editor@example.gov.vn"
          },
          "username": {
            "type": "string",
            "example": "editor01",
            "pattern": "^[a-z][a-z0-9._-]{2,63}$"
          },
          "displayName": {
            "type": "string",
            "example": "Biên tập viên 01",
            "minLength": 2,
            "maxLength": 200
          },
          "role": {
            "type": "string",
            "enum": [
              "system_admin",
              "editor",
              "reviewer",
              "publisher"
            ]
          },
          "delivery": {
            "type": "string",
            "enum": [
              "manual",
              "invite"
            ]
          },
          "temporaryPassword": {
            "type": "string",
            "writeOnly": true,
            "minLength": 12
          }
        },
        "required": [
          "email",
          "username",
          "displayName",
          "role",
          "delivery"
        ]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "reviewer@example.gov.vn"
          },
          "username": {
            "type": "string",
            "example": "reviewer01",
            "pattern": "^[a-z][a-z0-9._-]{2,63}$"
          },
          "displayName": {
            "type": "string",
            "example": "Kiểm duyệt viên 01",
            "minLength": 2,
            "maxLength": 200
          },
          "role": {
            "type": "string",
            "enum": [
              "system_admin",
              "editor",
              "reviewer",
              "publisher"
            ]
          },
          "expiresInHours": {
            "type": "integer",
            "default": 72,
            "minimum": 1,
            "maximum": 168
          }
        },
        "required": [
          "email",
          "username",
          "displayName",
          "role"
        ]
      },
...
    }
  }
}